Move web-component example to separate page
